requirements.txt
flask==0.11
flask-sqlalchemy==2.1
flask-migrate==2.0.0
flask-restful==0.2.1
mysql-python==1.2.5
gunicorn==19.6
python-dotenv==0.6.1
flask-httpauth==3.2.1
passlib==1.6.5
flask-login==0.4.0
flask-wtf==0.13.1
StringGenerator==0.1.9